requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
transformers
torch
fastapi
uvicorn
streamlit
requests